pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "refactor-framework"
version = "0.1.0"
description = "Framework for systematic legacy codebase refactoring with metrics tracking and reporting"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "radon>=6.0",
    "lizard>=1.17",
    "unidiff>=0.7",
    "jinja2>=3.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]
llm = [
    "anthropic>=0.39",
]

[project.scripts]
refactor-framework = "refactor_framework.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: tests requiring filesystem or subprocess",
    "slow: tests taking >5s",
]
